third_party/WebKit/LayoutTests/fast/box-decoration-break/box-decoration-break-parsing-expected.txt

Test to make sure the box-decoration-break property is correctly parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.style.getPropertyValue('-webkit-box-decoration-break') is 'slice'
PASS computedStyle.getPropertyValue('-webkit-box-decoration-break') is 'slice'
PASS e.style.getPropertyValue('-webkit-box-decoration-break') is 'clone'
PASS computedStyle.getPropertyValue('-webkit-box-decoration-break') is 'clone'
PASS e.style.getPropertyValue('-webkit-box-decoration-break') is ""
PASS computedStyle.getPropertyValue('-webkit-box-decoration-break') is 'slice'
PASS successfullyParsed is true

TEST COMPLETE
